Naga TypeInner::Pointer types are not TypeFlags::DATA.
Change `Validator::validate_type` not to mark pointers with the
`TypeFlags::Data` flag. The `TypeFlags::DATA` flag is supposed to correspond to
WGSL's concept of a 'plain type': a scalar, atomic, or composite type, excluding
textures, samplers, and pointers.

With this change, there is no type where `TypeFlags::DATA` and
`TypeFlags::INTERFACE` differ, but I'm not removing anything in this PR, because
I think we'll want to re-introduce a different distinction between them soon.